# Fix `@simd` for non 1 step CartesianPartition