@nanosoldier 2021-10-23
[Your benchmark job](https://github.com/JuliaLang/julia/pull/42736#issuecomment-949798349) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/3dfd4b6_vs_59aa3ed/report.md).