@vchuravy 2021-10-31
@nanosoldier `runbenchmarks("simd", vs=":master")`