@nanosoldier 2021-10-31
[Your benchmark job](https://github.com/JuliaLang/julia/pull/42736#issuecomment-955762656) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/3522071_vs_86f656d/report.md).