@N5N3 2021-11-01
Interesting results.
The first possible regression also appeared in [Nanosoldier Daily Job: 2021-10-28 vs 2021-10-26](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_date/2021-10/28/report.md) with a close time ratio (1.77 there and 1.76 here).
As for the second (and third, they are exactly the same), similar `possibe regression` on `("axpy!", "Float32", 4096)` could also be found in the above report.
So I think we can ingore them here.

BTW, is there any work remain to do on https://github.com/JuliaCI/BaseBenchmarks.jl/pull/284 @vchuravy