@N5N3 2021-11-21
Just notice that reused `simd_index` makes the position of `@inbounds` influence `@simd for`'s performance. (`BitArray`'s `broadcast` is the only use case I found in Base.)

On the other hand, I still think we'd better make `CartesianIndices`'s `simd_index` always inbound like before, since it is really strange that `@inbounds` influences the performance of index iteration.