@N5N3 2021-11-22
Rebasing solves Downloads error on FreeBSD.
The current `threads` error on Linux32 seems unrelated.
Just wonder whether further test is needed?
Or just re-run simd `Benchmark`? @vchuravy  @vtjnash